+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `isDelivered` function only considers output packages.
+
 ## [1.6.0] - 2021-03-17
 
 ### Added